+++ llm_guard/output_scanners/malicious_urls.py
@@ -2,11 +2,17 @@ import re
 from typing import List
 
 from llm_guard.transformers_helpers import pipeline_text_classification
-from llm_guard.util import logger
+from llm_guard.util import calculate_risk_score, logger
 
 from .base import Scanner
 
-_model_path = "elftsdmr/malware-url-detect"
+_model_path = "DunnBC22/codebert-base-Malicious_URLs"
+_model_path_onnx = "laiyer/codebert-base-Malicious_URLs-onnx"
+_malicious_labels = [
+    "defacement",
+    "phishing",
+    "malware",
+]
 
 # URL pattern
 url_pattern = re.compile(
@@ -19,25 +25,27 @@ class MaliciousURLs(Scanner):
     This scanner is used to scan and detect malicious URLs in the text.
 
     Large Language Models can occasionally produce malicious URLs in their outputs.
-    Using the "elftsdmr/malware-url-detect" model from HuggingFace, this class classifies
+    Using the "DunnBC22/codebert-base-Malicious_URLs" model from HuggingFace, this class classifies
     URLs as either malicious or benign to safeguard users from potential threats.
     """
 
-    def __init__(self, threshold=0.75, use_onnx: bool = False):
+    def __init__(self, threshold=0.5, use_onnx: bool = False):
         """
         Initializes an instance of the MaliciousURLs class.
 
         Parameters:
-            threshold (float): The threshold used to determine if the website is malicious. Defaults to 0.75.
+            threshold (float): The threshold used to determine if the website is malicious. Defaults to 0.5.
             use_onnx (bool): Whether to use the ONNX version of the model. Defaults to False.
         """
 
         self._threshold = threshold
         self._classifier = pipeline_text_classification(
             model=_model_path,
+            onnx_model=_model_path_onnx,
             truncation=True,
-            padding=True,
             use_onnx=use_onnx,
+            top_k=None,
+            max_length=512,
         )
 
     @staticmethod
@@ -53,21 +61,19 @@ class MaliciousURLs(Scanner):
             return output, True, 0.0
 
         logger.debug(f"Found {len(urls)} URLs in the output")
+        results = self._classifier(urls)
+        for url, result in zip(urls, results):
+            malicious_scores = [
+                item["score"] for item in result if item["label"] in _malicious_labels
+            ]
+            highest_malicious_score = max(malicious_scores)
+            if highest_malicious_score > self._threshold:
+                logger.warning(
+                    f"Detected malware URL '{url}' with score: {highest_malicious_score}"
+                )
 
-        urls_str = ", ".join(urls)
-        result = self._classifier(urls_str)
-        malware_score = (
-            result[0]["score"] if result[0]["label"] == "MALWARE" else 1 - result[0]["score"]
-        )
-        if malware_score > self._threshold:
-            logger.warning(
-                f"Detected malware URL with score: {malware_score}, threshold: {self._threshold}"
-            )
+                return output, False, calculate_risk_score(highest_malicious_score, self._threshold)
 
-            return output, False, round(malware_score, 2)
-
-        logger.debug(
-            f"Not malware URLs in the output. Max score: {malware_score}, threshold: {self._threshold}"
-        )
+        logger.debug(f"Not malware URLs in the output: {results}")
 
         return output, True, 0.0